requirements.txt
streamlit
python-dotenv
requests
trimesh
numpy
scipy
meshio
pyglet
